Update Master branch

+++ frontend/src/pages/Login/views/Login.vue
@@ -19,7 +19,7 @@
                 </button>
               </div>
               <p>Đăng nhập bằng email:</p>
-              <form>
+              <form v-on:submit.prevent="login()">
                 <div class="form-group">
                   <input
                     type="email"
@@ -28,10 +28,19 @@
                     placeholder="Email"
                     required
                     v-model="form.email"
+                    name="emailuser"
+                    data-vv-as="email"
+                    v-validate.continutes="'required|email'"
                   />
                   <button class="btn icon">
                     <i class="material-icons">mail_outline</i>
                   </button>
+                </div>
+                <div class="form-group">
+                  <span
+                    style="color:red;"
+                    v-if="errors.has('emailuser')"
+                  >{{ errors.first('emailuser') }}</span>
                 </div>
                 <div class="form-group">
                   <input
@@ -41,16 +50,36 @@
                     placeholder="Mật khẩu"
                     required
                     v-model="form.pass"
+                    name="inputPassword"
+                    data-vv-as="password"
+                    v-validate="'required'"
                   />
                   <button class="btn icon">
                     <i class="material-icons">lock_outline</i>
                   </button>
                 </div>
                 <div class="form-group">
+                  <span
+                    style="color:red;"
+                    v-if="errors.has('inputPassword')"
+                  >{{ errors.first('inputPassword') }}</span>
+                </div>
+                <div class="form-group">
                   <p style="color: red;">{{login_error}}</p>
                 </div>
-                <button type="button" class="btn button" v-if="!disbleLogin" v-on:click="login()">Đăng nhập</button>
-                <button type="button" class="btn button" v-if="disbleLogin" v-on:click="login()" disabled>Đăng nhập</button>
+                <button
+                  type="button"
+                  class="btn button"
+                  v-if="!disbleLogin"
+                  v-on:click="login()"
+                >Đăng nhập</button>
+                <button
+                  type="button"
+                  class="btn button"
+                  v-if="disbleLogin"
+                  v-on:click="login()"
+                  disabled
+                >Đăng nhập</button>
                 <div class="callout">
                   <span>
                     Chưa có tài khoản?
@@ -85,6 +114,8 @@ import register_router from "../../Register/router";
 import home_router from "../../Home/router";
 import { TokenService } from "../../../services/storage.service";
 
+import VeeValidate from "vee-validate";
+
 export default {
   name: "login",
   data() {
@@ -107,24 +138,33 @@ export default {
     },
     login: function() {
       this.disbleLogin = true;
-      this.$http
-        .post(this.LOGIN_CONTROLLER + "/login", { user: this.form })
-        .then(val => {
-          if (val.data.success) {
-            this.login_error = null;
-            this.user_info = val.data.content.data.user;
-            TokenService.removeToken();
-            TokenService.removeRefreshToken();
-            TokenService.saveToken(val.data.content.data.token);
-            //TokenService.saveRefreshToken(response.data.refresh_token);
-            home_router.push({ name: "home" });
-            location.reload();
-          } else {
-            this.login_error = "Email hoặc mật khẩu không đúng";
-            this.disbleLogin = false;
+      this.$validator
+        .validateAll()
+        .then(res => {
+          if (res) {
+            this.$http
+              .post(this.LOGIN_CONTROLLER + "/login", { user: this.form })
+              .then(val => {
+                if (val.data.success) {
+                  this.login_error = null;
+                  this.user_info = val.data.content.data.user;
+                  TokenService.removeToken();
+                  TokenService.removeRefreshToken();
+                  TokenService.saveToken(val.data.content.data.token);
+                  //TokenService.saveRefreshToken(response.data.refresh_token);
+                  home_router.push({ name: "home" });
+                  location.reload();
+                } else {
+                  this.login_error = "Email hoặc mật khẩu không đúng";
+                  this.disbleLogin = false;
+                }
+              })
+              .catch(err => {});
           }
         })
-        .catch(err => {});
+        .catch(err => {
+          console.log(err);
+        });
     }
   }
 };
